+++ src/main/java/platform/controller/RestUserController.java
@@ -4,7 +4,10 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import liquibase.util.StringUtil;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -14,6 +17,7 @@ import platform.domain.User;
 import platform.service.RoleService;
 import platform.service.UserService;
 
+import java.util.Collection;
 import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -33,6 +37,10 @@ public class RestUserController implements UserRestApi {
     @Override
     @Operation(summary = "Add new user")
     public ResponseEntity<String> add(UserVO body) {
+        Collection<GrantedAuthority> authorities = (Collection<GrantedAuthority>) SecurityContextHolder.getContext().getAuthentication().getAuthorities();
+        if (!userService.hasRole("ROLE_ADMIN", authorities)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(getJSONMsg(403,"error", "Forbidden", "User do not have permissions"));
+        }
         String username = body.getUsername();
         String password = body.getPassword();
         String email = body.getEmail();
@@ -121,6 +129,10 @@ public class RestUserController implements UserRestApi {
     @Override
     @Operation(summary = "Restore user by id")
     public ResponseEntity<String> restore(Integer id) {
+        Collection<GrantedAuthority> authorities = (Collection<GrantedAuthority>) SecurityContextHolder.getContext().getAuthentication().getAuthorities();
+        if (!userService.hasRole("ROLE_ADMIN", authorities)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(getJSONMsg(403,"error", "Forbidden", "User do not have permissions"));
+        }
         User user = userService.findById(id);
         if (user != null) {
             int enabled = user.getEnabled();
@@ -137,6 +149,10 @@ public class RestUserController implements UserRestApi {
     @Override
     @Operation(summary = "Update user by id")
     public ResponseEntity<String> update(Integer id, UserVO body) {
+        Collection<GrantedAuthority> authorities = (Collection<GrantedAuthority>) SecurityContextHolder.getContext().getAuthentication().getAuthorities();
+        if (!userService.hasRole("ROLE_ADMIN", authorities)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(getJSONMsg(403,"error", "Forbidden", "User do not have permissions"));
+        }
         String username = body.getUsername();
         String password = body.getPassword();
         String email = body.getEmail();
@@ -186,6 +202,10 @@ public class RestUserController implements UserRestApi {
     @Override
     @Operation(summary = "Delete user by id")
     public ResponseEntity<String> delete(Integer id) {
+        Collection<GrantedAuthority> authorities = (Collection<GrantedAuthority>) SecurityContextHolder.getContext().getAuthentication().getAuthorities();
+        if (!userService.hasRole("ROLE_ADMIN", authorities)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(getJSONMsg(403,"error", "Forbidden", "User do not have permissions"));
+        }
         if (userService.remove(id)) {
             return ResponseEntity.ok(getJSONMsg(200,"success", "OK","User with id '" + id + "' successfully deleted"));
         } else {
